+++ scripts/deploy.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+#
+# Copy the artifacts for a state from fileout to docs
+#
+# For example:
+#
+# scripts/deploy.sh NC
+
+XX=$1
+FROM_DIR=../../iCloud/fileout
+TO_DIR=docs
+
+cp $FROM_DIR/images/$XX*.svg $TO_DIR/assets/images
+cp $FROM_DIR/_data/$XX*.csv $TO_DIR/_data